refactor(ts): service index file
